Replace MOS 6502 components with synthesizable versions
Removed _dsl suffix from all component files and class names:
- alu.rb: Full ALU with BCD arithmetic support
- memory.rb: 64KB memory with RAM/ROM regions
- registers.rb: Registers, StackPointer, ProgramCounter, latches
- status_register.rb: P register with all flags
- address_gen.rb: Address generator for all 13 modes
- instruction_decoder.rb: 151-opcode decode table
- control_unit.rb: 25+ state FSM for instruction sequencing

All components include to_verilog class methods generating synthesizable Verilog.
Added backward compatibility aliases (StackPointer6502, ProgramCounter6502).
Tests: 191 MOS 6502 tests, 372 core RHDL tests all passing.
